CLOUDERA-BUILD add superuser_acl/user_acl service-level parameters to CSD
In CDH 5.11, these parameters can only be used for username-based access
controls; group-based controls will be added in a future release. For the
time being, attempts to add group names will fail in validation.

The regex used to parse the ACLs is a monster. In English: "an ACL parameter
may be set to the empty string, a '*', or a comma-separated list of entries.
Each entry may begin with any character except for a forbidden set (some
punctuation and whitepsace), and then followed by any other character(s)
except a comma (as the comma delimits the entries).

Change-Id: I546580565fde967f5dd120b74b5ab4e0247d9c93