CLOUDERA-BUILD remove RPC tri-states from CSD
We've decided that this is probably too complex for the majority of CM
users. Moreover, CM's "Enable Kerberos" wizard isn't going to set the
tri-states on our behalf.

So, we're reverting back to a single checkbox to control all Kerberos usage,
and for that checkbox to turn the various knobs within Kudu to require
Kerberos-based authentication and encryption.

To be clear: Kudu itself retains the tri-state parameters, it's just that
we'll map this single CM-only security parameter into them.

Change-Id: I327ad238ef4e1c48bb3bbea40a61cd9727a71066