CLOUDERA-BUILD Clean up after parcel upgrade
Considering the size of parcels, it may be desirable to remove the
old parcel distributions and downloads after upgrading. This patch
adds a flag to the upgrade script indicating whether or not to clear
unused parcels (off by default).

A parcel is considered "unused" if its status is DISTRIBUTED or
DOWNLOADED after the upgrade is complete, and will be set to
AVAILABLE_REMOTELY.