CLOUDERA-BUILD Wait for stop() before cleaning
The Jenkins upgrade script calls service.restart() and immediately
proceeds to remove old parcels. This can result in an issue where
the old parcel being used by the service is removed before the service
is completely stopped on all servers, preventing the restart() call
from completing.

This patch remediates this by explicitly using the service functions
stop() and start() instead of restart(), and waiting for the service
to be in the "STOPPED" state before starting again.