[csd] Add softMin for memory_hard_limit_bytes
One of the most common user gotchas with Kudu is that CM often
configures Kudu tablet servers with 1GiB of memory, which turns
out to be too little. This is CM working as intended: the worker
nodes' memory is oversubscribed and therefore CM configures new
services with the minimum memory value. Frequently, the
oversubscription does not hold in reality, and CM's
configuration means users see bad performance. This change adds
a softMin of 4GiB to memory_limit_hard_bytes (the current
default), so that users will see a warning when the configured
value falls below this. This should help alert users that the
default configuration likely ought to be changed.